remove unnecessary renames
