+++ commandline/src/main/java/tech/pegasys/web3signer/commandline/subcommands/Eth1SubCommand.java
@@ -18,8 +18,8 @@ import static tech.pegasys.web3signer.commandline.DefaultCommandValues.PATH_FORM
 import static tech.pegasys.web3signer.commandline.DefaultCommandValues.PORT_FORMAT_HELP;
 import static tech.pegasys.web3signer.commandline.util.RequiredOptionsUtil.checkIfRequiredOptionsAreInitialized;
 
+import tech.pegasys.web3signer.commandline.PicoCliAwsKmsParameters;
 import tech.pegasys.web3signer.commandline.PicoCliEth1AzureKeyVaultParameters;
-import tech.pegasys.web3signer.commandline.PicoCliKmsAwsParameters;
 import tech.pegasys.web3signer.commandline.annotations.RequiredOption;
 import tech.pegasys.web3signer.commandline.config.client.PicoCliClientTlsOptions;
 import tech.pegasys.web3signer.core.Eth1Runner;
@@ -148,7 +148,7 @@ public class Eth1SubCommand extends ModeSubCommand implements Eth1Config {
 
   @CommandLine.Mixin private PicoCliEth1AzureKeyVaultParameters azureKeyVaultParameters;
 
-  @CommandLine.Mixin private PicoCliKmsAwsParameters awsParameters;
+  @CommandLine.Mixin private PicoCliAwsKmsParameters awsParameters;
 
   @Override
   public Runner createRunner() {
